test_pdf_generation.py
#!/usr/bin/env python
"""Test script to verify advanced PDF generation with photo support"""
import os
import django
from PIL import Image as PILImage

os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'eye_detection.settings')
django.setup()

from detection.models import Patient, Detection
from utils.pdf_generator import generate_pdf
from django.conf import settings

print("\n" + "="*70)
print("ADVANCED PDF REPORT GENERATION TEST")
print("="*70)

# Create a test patient
patient, _ = Patient.objects.get_or_create(
    name="Test Patient",
    defaults={'age': 42, 'gender': 'M', 'phone': '+91-9876543210'}
)
print(f"\n[1] Patient created: {patient.name} (ID: {patient.patient_id})")

# Create a dummy test image (colored square)
test_image_path = "media/uploads/test_image.png"
os.makedirs(os.path.dirname(test_image_path), exist_ok=True)
img = PILImage.new('RGB', (224, 224), color='blue')
img.save(test_image_path, 'PNG')
print(f"[2] Test image created: {test_image_path}")

# Create a detection record
detection, created = Detection.objects.get_or_create(
    detection_id="TEST001",
    defaults={
        'patient': patient,
        'image': 'uploads/test_image.png',
        'predicted_disease': 'glaucoma',
        'confidence_score': 85.5,
        'severity': 'SEVERE',
        'english_explanation': "Glaucoma is a group of eye diseases that damage the optic nerve due to elevated intraocular pressure. Early treatment can halt progression.",
        'tamil_explanation': "கண் அழுத்த நோய் என்பது கண்ணினுள் அதிகரித்த அழுத்தம் பார்வை நரம்பை பாதிக்கும் ஒரு கண் நோய். ஆரம்பகால சிகிச்சை முக்கியம்.",
        'symptoms': "Gradual loss of peripheral vision\nTunnel vision in advanced stages\nSevere eye pain\nHeadache and nausea\nHalos around lights",
        'causes': "Elevated intraocular pressure\nFamily history\nAge above 60\nThin cornea\nSteroid use",
        'treatment': "Eye drop medications\nLaser trabeculoplasty\nGlaucoma surgery\nRegular monitoring",
        'prevention': "Regular eye exams after age 40\nKnow family history\nExercise regularly\nProtect eyes from injury\nTake medications consistently",
        'all_probabilities': '{"glaucoma": 85.5, "cataract": 8.2, "diabetic_retinopathy": 4.1, "normal": 2.2}',
    }
)
print(f"[3] Detection created: {detection.detection_id}")
print(f"    Disease: {detection.predicted_disease.title()}")
print(f"    Confidence: {detection.confidence_score}%")
print(f"    Severity: {detection.severity}")

# Generate PDF
print(f"\n[4] Generating advanced PDF report...")
try:
    pdf_path = generate_pdf(detection)
    print(f"    ✅ PDF generated successfully")
    print(f"    Location: {pdf_path}")
    
    if os.path.exists(pdf_path):
        size_mb = os.path.getsize(pdf_path) / (1024*1024)
        print(f"    File size: {size_mb:.2f} MB")
    
    detection.report_pdf = os.path.relpath(pdf_path, 'media/')
    detection.save()
    print(f"    ✅ Report saved to database")
    
except Exception as e:
    print(f"    ❌ PDF generation failed: {e}")
    import traceback
    traceback.print_exc()

print("\n" + "="*70)
print("TEST COMPLETE")
print("="*70)
print("\nFeatures tested:")
print("✓ Patient photo embedding")
print("✓ Bilingual support (English + Tamil)")
print("✓ Confidence metrics and probability breakdown")
print("✓ Advanced medical formatting")
print("✓ Severity indicators")
print("✓ Clinical recommendations")
print("\n")
