requirements.txt
# Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0

# Deep Learning
tensorflow==2.15.0
numpy==1.24.3
scikit-learn==1.3.2
matplotlib==3.8.2

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# PDF
reportlab==4.0.7

# AI
openai==1.3.7
python-decouple==3.8

# Production
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
